Chapter05/requirements.txt
-i https://pypi.org/simple
certifi==2019.9.11
chardet==3.0.4
idna==2.8
python-dateutil==2.8.1
requests==2.22.0
six==1.13.0
urllib3==1.25.6
